+++ LoipvRemoteTests/UI/Forms/UnifiedWindowLayoutTests.cs
@@ -7,6 +7,20 @@ namespace LoipvRemoteTests.UI.Forms
     [TestFixture]
     public class UnifiedWindowLayoutTests
     {
+        [TestCase(44, 35)]
+        [TestCase(58, 46)]
+        public void TopMenuHeightIsReducedByTwentyPercent(int originalHeight, int expectedHeight)
+        {
+            Assert.That(UnifiedWindowHeader.ReduceHeaderHeight(originalHeight), Is.EqualTo(expectedHeight));
+        }
+
+        [TestCase(35, 400)]
+        [TestCase(58, 464)]
+        public void MenuHostKeepsAllMainMenuItemsVisible(int headerHeight, int expectedWidth)
+        {
+            Assert.That(UnifiedWindowHeader.MenuHostWidthFor(headerHeight), Is.EqualTo(expectedWidth));
+        }
+
         [Test]
         public void ContentBeginsBelowVisibleHeader()
         {